backend/requirements.txt
# ──────────────────────────────────────────────
# Core Django + DRF
# ──────────────────────────────────────────────
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
drf-spectacular>=0.27

# ──────────────────────────────────────────────
# Database
# ──────────────────────────────────────────────
psycopg2-binary>=2.9.9
pgvector>=0.3.0

# ──────────────────────────────────────────────
# Cache / Broker
# ──────────────────────────────────────────────
redis>=5.0
django-redis>=5.4

# ──────────────────────────────────────────────
# Async / WebSockets (Django Channels)
# ──────────────────────────────────────────────
channels[daphne]>=4.0
channels-redis>=4.2
daphne>=4.0

# ──────────────────────────────────────────────
# Celery (async task queue)
# ──────────────────────────────────────────────
celery[redis]>=5.3
django-celery-beat>=2.6
django-celery-results>=2.5

# ──────────────────────────────────────────────
# HTTP client
# ──────────────────────────────────────────────
httpx>=0.27
aiohttp>=3.9

# ──────────────────────────────────────────────
# AI / ML / LLM
# ──────────────────────────────────────────────
openai>=1.30
langgraph>=0.2
langchain-openai>=0.1
langchain-core>=0.2
sentence-transformers>=3.0
FlagEmbedding>=1.2

# ──────────────────────────────────────────────
# Authentication – Clerk JWT
# ──────────────────────────────────────────────
python-jose[cryptography]>=3.3
cryptography>=41.0
svix>=1.0

# ──────────────────────────────────────────────
# External services
# ──────────────────────────────────────────────
tavily-python>=0.3
weasyprint>=62.0

# ──────────────────────────────────────────────
# Utilities
# ──────────────────────────────────────────────
python-dotenv>=1.0
django-environ>=0.11
Pillow>=10.0
numpy>=1.26
django-extensions>=3.2
gunicorn>=21.0
uvicorn>=0.29

# ──────────────────────────────────────────────
# Type stubs / dev tools (optional)
# ──────────────────────────────────────────────
ipython>=8.24
